requirements.txt
numpy
pandas
opencv-python
matplotlib
scikit-learn
scikit-image
streamlit
joblib
pillow
